Switch sensor value pipeline from float to double
- sensors.h/sensors.c: sensor_data_t.value, sensors_update(),
  sensors_update_with_link(), and sensors_get_value() now carry double
  instead of float, since this is the shared choke point every sensor
  reading passes through.
- weight.c, rcwl9620.c, a02yyuw.c, ezo_ph.c, pump.c, flow_sensor.c:
  driver-side readings and running totals converted to double to match
  (including switching sscanf("%f", ...) parses to "%lf" where values
  come from ASCII responses).
- settings.h/settings.c: flow_sensor_total_liters and
  settings_save_flow_total() are now double; added the same
  legacy-float NVS blob migration used for flow_sensor_liters_per_pulse
  so an existing 4-byte flow_tot%d blob upconverts instead of a
  size-mismatched read leaving garbage bytes.
- DS18B20 temperature and BTHome-derived readings intentionally left as
  float, since both are bound to external managed components
  (espressif/ds18b20, jcodybaker/bthome) whose APIs return float.

+++ main/a02yyuw.c
@@ -62,7 +62,7 @@ static void a02yyuw_task(void *pvParameters) {
             continue;
         }
 
-        float distance_cm = distance_mm / 10.0f;
+        double distance_cm = distance_mm / 10.0;
         ESP_LOGD(TAG, "Distance: %.1f cm", distance_cm);
         sensors_update(sensor_id_cm, distance_cm, true);
     }
